Cache busting for CSS
This updates the pandoc command that generates the html to use
'sha256sum' to append the first 16 characters to the stylesheets to
hopefully have better control over caching.

Additionally, the CI workflow is updated to include steps for setting
cache headers on the content uploaded to S3. The site is also updated to
explicitly include the fontawesome stylesheets that are needed, rather
than 'all'.